edit json, sync v0.1 -> 0.2, load/save player info
