+++ flip_world.h
@@ -12,6 +12,7 @@
 #include <dialogs/dialogs.h>
 
 #define TAG "FlipWorld"
+#define VERSION 0.2
 #define VERSION_TAG "FlipWorld v0.2"
 
 // Define the submenu items for our FlipWorld application
