+++ game/level.c
@@ -1,5 +1,5 @@
 #include <game/level.h>
-
+#include <flip_storage/storage.h>
 static void level_start(Level *level, GameManager *manager, void *context)
 {
     UNUSED(manager);
